refactor: outsourced api call functions in a projects.api.ts and unwrapBody in a central helper
